Split workspace package route tests
